refactor: env-var overrides, error handling, and cleanup for testability
- app.js: env-var path overrides (LEADERBOARD_CONFIG_PATH, LEADERBOARD_DATA_PATH,
  LEADERBOARD_LOG_PATH, LEADERBOARD_ADMINDATA_PATH, LEADERBOARD_CONFIG_BACKUP_PATH,
  LEADERBOARD_PORT), LEADERBOARD_SKIP_REFRESH to disable child process, pass res
  to Util.post calls, fix hardcoded ./admindata.json in login handler, export server
- Util.js: post(req, res, callback) — respond 400 on malformed JSON instead of
  silently hanging, move findContributor from app.js
- API.js: remove process.exit() on bad credentials — log error and return instead,
  support LEADERBOARD_CONFIG_PATH env var
- Remove dead .dockerignore (Docker support removed previously)
- Remove stale tests/access.sh and tests/api-call-tests.sh (brute-force scripts
  hitting dead external URLs)
